pyproject.toml
[tool.poetry]
name = "dpd-test"
version = "0.1.0"
description = ""
authors = ["ferril <ferril.darkdiver@gmailc.om>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
redis = "^3.3"
djangorestframework = "^3.10"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
